initial work on generating JSON (de)serialization + .ts types from it
so that we can get better type information than simply `any` in the
.ts/flow type annotations.

This is a multi-phase process:
1) Generate (using `schemars` traits) .json schemas for all types. This
is done by running the program in rust/json-gen which will export these
to rust/json-gen/schemas
2) Running scripts/run-json2ts.js to convert these to .ts files
3) We will need to convert the names to add in the JSON suffix to these
(not implemented yet)
4) Running scripts/json-ts-types.js to replace any `any` return types
with the correct *JSON type generated above

some files are hardcoded so they don't line up right now and some steps
aren't working as the schemas generated in 1) inline ALL referenced
schemas within a single schema so in step 2) we end up with a lot of
duplicate type definitions. We might be able to just have a stage to
remove duplicates from the schemas, or maybe there's a way to change the
json to ts tool to not do that.

Some types also throw runtime errors when serializing e.g. non-string
keys despite compiling fine since serde != serde_json