PointerAddress support for >u32::MAX (#33)
* PointerAddress support for >u32::MAX

Fixes #364 in a way that requires the least changes.

Only users who are manually creating `Pointer` to make a new
`PointerAddress` from parts will be affected by this API change.

We could also alternatively change *ALL* references of `Slot` and `TransactionIndex`
to support `BigNum` instead of `u32`. This would be inline with the
alonzo.cddl binary spec, but would make use of the library more
annoying, especially from javascript/wasm, as well as requiring more code changes.

I don't see why you would have a transaction index greater than even a
single byte, let alone 2^32 - 1. But since that's how the binary spec
is, it's possible we'll run into on-chain data that do this in the future.

This isn't the only place where we run into issues like this  as there are
edge cases anywhere where `nint` is specified since the CBOR library we
support doesn't handle numbers between `i64::MIN` and `-u64::MAX`.

* variable_nat_decode() check for overflow / add test case

Co-authored-by: rooooooooob <raftias@gmail.com>