Slot/TransactionIndex/CertificateIndex changes (#36)
When `Slot` was changed to `BigNum`, it wasn't updated for wasm-suitable
API in every location (still taken by value in some spots), so this
includes fixes for that.

Also updates the changes in #33 to work with the global `Slot` change
that was done prior.

`TransactionIndex` and `CertificateIndex` were globablly changed to
`BigNum` for consistency and because we'll likely run into those
on-chain if they are already there inside of pointer addresses.

There still remains other uses of `u32`, especially in lower priority
structures like block headers and updates. Those should be fine for now
as we might end up removing `BigNum` soon anyway, and those are not as
succeptible to random users posting them on-chain.

Warning in tests were fixed as well.