Workspaces split (#153)
* Workspace refactor initial moving existing code

* Move CIP25 library into CML.

Existed before at https://github.com/dcSpark/cip25-lib

* Base code changes for module split

Going on the previous two commits bringing in exisitng code over to the
workspace structure, this makes it work together using each other as
dependencies and such + misc fixes so it compiles. This includes the
CIP25 crate as well.

* Move on-chain crypto into chain crate

All things specified in babbage/crypto.cddl are now back in the chain
crate.

To work around some things e.g. signatures, hashes being uses in the
crypto crate we have those being their base locations with chain/crypto
having a `ChainCrypto<T>` wrapper around this.

We have the to/from raw bytes in a `RawBytesEncoding` trait to help out
as well for when things are raw bytes.

All crypto bindings are still in the `wasm` crate. We will need to look
into a separate `crypto-wasm` crate and first test out how those would
interact and if it's possible/a good idea.

* crypto-wasm crate added

Crypto wasm bindings moved into separate `crypto-wasm` crate. The
on-chain representations are still in `wasm`.