CIP-25 build script (#160)
* CIP-25 build script

CIP-25 was used as an experiment to try out the new split crate
architecture. The build scripts were adapted to work for this new
format. The old build scripts are moved to `/scripts/legacy/`. This
should allow us to have per wasm target npm crates as well.
If this works out, we'll adapt the other WASM crates to use this build
structure too.

* Fix package scripts

Co-authored-by: Sebastien Guillemot <sebastiengllmt@gmail.com>