Flaky random-improve tests now try 3 times (#174)
Before this change I ran the tests 100 times and had 3 failures (i.e. 1
or more of those 5 tests failed).

This can occasionally be annoying if it causes the CI to report a
failure.

After this change each one runs 3 times which should make it extremely
unlikely to have a test fail as the same test would have to fail 3
times in a row to fail. Using those stats from above that's a <0.0027%
(assuming each test 3% chance - likely much lower as that's the chance
for only 1 or more of the 5 tests failing) chance of failure.