CIP25 ignore properly extra data EVERYWHERE (#180)
Includes a fix for test case in #175

This also includes fixes for extra data in tables too, not just those
ones in map-structs.

A test case is added that covers basically anywhere you could add extra
data, but under the assumption that keys of the same CBOR type as
somewhere in the spec must conform to the spec. The rest of the data is
ignored properly and can be whatever. Also, constants e.g. string
members are also allowed that don't conform and are ignored (this is
what #175 tested).

It could be worth adding this operation to cddl-codegen if we get more
metadata-based specs although we would need to do it a bit more
generically since the other data wouldn't necessarily be a metadatum
(unless we hardcode CML-related types into cddl-codegen but I'd rather
not do that).: