BigInt + bounded bytes
bounded-bytes logic including adapting for preserving encodings.
This is 100% an encoding detail as we can express the chunking logic in
the encoding variable and expose just a simple bytes type.

bigint hand-written struct to cover the specific big integer encoding
(which can be `Int` too), and also is chunked i.e. bounded bytes