WASM build fixes + lint fixes
WASM build fixes: duplicate name between chain-wasm and multi-era-wasm
which does not work when exporting to typescript.

WASM build fixes: few misc things not included in previous commit

lint fixes: mostly unused vars/imports/etc but also Certificate is too
big which caused issues on its own + use in Result as Err variant.

At some point we could offer a way to have cddl-codegen Box variants if
there's a large variance and fix it that way but until then I just
allow()'d it.

GovAction::I6 -> GovAction::InfoAction renamed as well. this is how it
should be according to the .cddl spec but cddl-codegen ignores the @name
comment in this specific place: ddcSpark/cddl-codegen#211
so this was resolved by hand afterwards.