Prefix CIP25/CIP36 types
Many CIP25/CIP36 types were pretty confusing when used from WASM as
there was no module scoping so everything was put together in the same
scope as chain/etc so things like "Delegation" were not obviously CIP36
for example.

We now prefix all types in both crates with CIP25/CIP36 to help with
this.