Add in missing [package] fields to crates' Cargo.tomls. (#285)
This results in a warning upon `cargo package` / `cargo publish` and
these were present in the old rust/Cargo.toml.

Points to the root readme/docs. We might want crate-specific ones later
on but right now the crate-specific readmes for some of the crates are
rather barebones on their own compared to the more helpful root readme.