Add missing hashing/crypto WASM bindings
Fixes #289

For script data hashing we do not pass in the witness set encodings like
in the rust bindings. We might wish to expose the CBOR encodings
generally, or for this API we could have something that takes in the
entire witness set. This function is likely to be used more indirectly
via the builders anyway.