Plutus Datum JSON to use Cardano Node format (#311)
* Plutus Datum JSON to use Cardano Node format

To replace the redundant auto-generated plutus datum JSON format with a
standardized one even when used from within other structures.

* Move metadata.rs from cml-core > cml-chain/auxdata

Having this in cml-core/cml-core-wasm was causing issues implementing
traits on the types and it is disjoint from the rest of the on-chain
types.

It was originally moved to cml-core with the idea of metadata-specific
libraries (e.g. cip25, cip36) could use this without needing to import
the much, much heavier cml-chain (+cml-crypto). In the end both of those
libraries ened up pulling in cml-chain in order to hook into things like
`PolicyId` and `AssetName`.

* Tx Metadatum JSON using cardano-node for serde impls